Fix handling of Pause and Manual mode
Previously, once you toggled `pause`, it also modified the `manual` flag. This meant that if you paused and restarted, automatic playback could not resume as it left the `manual` flag set to `true`. You had to pause/unpause first to resume automatic playback. Now the `manual` flag acts as the initial value for `pause` on start/restart which makes the behaviour more predictable.